Moved a method from the EntityPet class to the EntityBase class
