+++ NMS_Master/src/main/java/simplepets/brainsynder/nms/entity/EntityPet.java
@@ -3,12 +3,7 @@ package simplepets.brainsynder.nms.entity;
 import lib.brainsynder.nbt.StorageTagCompound;
 import lib.brainsynder.sounds.SoundMaker;
 import lib.brainsynder.utils.Colorize;
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.Registry;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.protocol.Packet;
-import net.minecraft.network.protocol.game.ClientboundAddEntityPacket;
-import net.minecraft.network.protocol.game.ClientboundAddMobPacket;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EntityType;
@@ -537,20 +532,6 @@ public abstract class EntityPet extends EntityBase implements IEntityPet {
     @Override
     public EntityType<?> getType() {
         return entityType;
-    }
-
-    @Override
-    public Packet<?> getAddEntityPacket() {
-        try {
-            ClientboundAddMobPacket packet = new ClientboundAddMobPacket(this);
-            Field type = packet.getClass().getDeclaredField("c");
-            type.setAccessible(true);
-            type.set(packet, Registry.ENTITY_TYPE.getId(originalEntityType));
-            return packet;
-        } catch (Exception ex) {
-            ex.printStackTrace();
-        }
-        return new ClientboundAddEntityPacket(this, originalEntityType, 0, new BlockPos(getX(), getY(), getZ()));
     }
 
     private void glowHandler(Player player, boolean glow) {
